+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Default governor choice matched to the current cpufreq driver.
+  Results in better support for non-intel processors.
+- Make use of `scaling_available_frequencies` to limit choice (of frequencies)
+
 ### Fixed
 - State switching issue (failure to read power supply online value)